api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.1",
    "azure-ai-textanalytics>=5.3.0",
    "azure-ai-translation-text>=1.0.1",
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.6.0",
    "cohere>=5.19.0",
    "fastapi-cli>=0.0.14",
    "fastapi[cli]>=0.119.1",
    "openai>=2.6.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]
